dd
aa

+++ cat/script.js
@@ -82,9 +82,22 @@ function store(){
 }
 
 function save(){
-  localStorage.setItem("pop", num);
-  localStorage.setItem("pScore", plusScore);
-  alert(num+'회 저장 완료! 로컬 기반으로 캐쉬 삭제 시 초기화 됩니다.');
+    if (!confirm("50% 확률로 게임을 저장 할 수 있습니다. 실패 시 모든 수치가 초기화됩니다.")) {
+      
+    } else {
+      if(Math.floor(Math.random() * 100) <= 50){
+          localStorage.setItem("pop", num);
+          localStorage.setItem("pScore", plusScore);
+         alert(num+'회 저장 완료! 로컬 기반으로 캐쉬 삭제 시 초기화 됩니다.');
+      }else{
+         alert('저장 실패!!! 모든 값이 초기화 되었습니다.');
+         plusScore =1;
+         num=0;
+         update();
+      }
+
+  
+}
 }
 function load(){  
   var SaveData = localStorage.getItem("pop");
